requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
influxdb==5.2.3
paho-mqtt==1.5.0
pyaml==20.3.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
